+++ breeze-core/src/test/kotlin/com/windea/breezeframework/core/extensions/StringExtensionsTest.kt
@@ -6,7 +6,7 @@ import com.windea.breezeframework.core.enums.text.ReferenceCase.*
 import kotlin.test.*
 
 class StringExtensionsTest {
-	@Test //TESTED
+	@Test
 	fun letterCaseTest() {
 		assertEquals("abcAbcAbc", "AbcAbcAbc".switchCaseBy(camelCase))
 		assertEquals("AbcAbcAbc", "abcAbcAbc".switchCaseBy(PascalCase))
@@ -19,7 +19,7 @@ class StringExtensionsTest {
 		assertEquals("ABCABC", "ABcABc".switchCaseBy(UPPERCASE))
 	}
 
-	@Test //TESTED ALMOST
+	@Test ALMOST
 	fun referenceCaseTest() {
 		assertEquals("Abc.Abc", "Abc.Abc".switchCaseBy(JavaReference))
 		assertEquals("abc.abc[1].abc", "/abc/abc/1/abc".switchCaseBy(JavaReference))
@@ -27,25 +27,25 @@ class StringExtensionsTest {
 		assertEquals("/abc/abc/1/abc", "/abc/abc/1/abc".switchCaseBy(PathReference))
 	}
 
-	@Test //TESTED
+	@Test
 	fun repeatExtensionTest() {
 		println("abc".repeat(3))
 		println("abc".flatRepeat(3))
 	}
 
-	@Test //TESTED
+	@Test
 	fun replaceExtensionTest() {
 		val string = "abc123abc123"
 		println(string.replace("""([a-z])\d""".toRegex(), "$1"))
 		println(string.replaceLooped("""([a-z])\d""".toRegex(), "$1"))
 	}
 
-	@Test //TESTED
+	@Test
 	fun messageFormatTest() {
 		assertEquals("123a123b123", "123{0}123{1}123".messageFormat("a", "b"))
 	}
 
-	@Test //TESTED
+	@Test
 	fun customFormatTest() {
 		assertEquals("1a2b3", "1{}2{}3".customFormat("{}", "a", "b"))
 		assertEquals("1a2b3", "1\${}2\${}3".customFormat("\${}", "a", "b"))
@@ -55,13 +55,13 @@ class StringExtensionsTest {
 		assertEquals("1b2a3", "1{bbb}2{aaa}3".customFormat("{name}", "aaa" to "a", "bbb" to "b"))
 	}
 
-	@Test //TESTED
+	@Test
 	fun quoteTest() {
 		assertEquals(""""'1\"2'"""", """'1"2'""".quote('"'))
 		assertEquals("""'1"2'""", """"'1\"2'"""".unquote())
 	}
 
-	@Test //TESTED
+	@Test
 	fun transformInTest() {
 		"\\Q abc abc \\E..\\Q abc abc abc \\E abc abc \\Q abc abc 123 \\E".transformIn("\\Q", "\\E") {
 			it.trim().split(" ").joinToString(", ", "{", "}") { s -> s.dropLast(1) }
@@ -84,6 +84,7 @@ class StringExtensionsTest {
 
 		assertTrue("/abc/123/def" matches "/abc/-/def".toRegexBy(MatchType.PathReference))
 		assertTrue("/abc/123/def" matches "/abc/[]/def".toRegexBy(MatchType.PathReference))
+		assertTrue("/abc/123/def" matches "/abc/[23-234]/def".toRegexBy(MatchType.PathReference))
 		assertTrue("/abc/123/def" matches "/abc/[b]/def".toRegexBy(MatchType.PathReference))
 		assertTrue("/abc/123/def" matches "/abc/[b]/def".toRegexBy(MatchType.PathReference))
 		assertTrue("/abc/123/def" matches "/{}/123/def".toRegexBy(MatchType.PathReference))