+++ breeze-dsl/breeze-dsl-critic-markup/src/test/kotlin/com/windea/breezeframework/dsl/criticmarkup/CriticMarkupTextDslTest.kt
@@ -3,7 +3,7 @@ package com.windea.breezeframework.dsl.criticmarkup
 import org.junit.*
 
 class CriticMarkupTextDslTest {
-	@Test //TESTED SO EASY
+	@Test SO EASY
 	fun test2() {
 		criticMarkupText {
 			"""