+++ breeze-dsl/breeze-dsl-json/src/test/kotlin/com/windea/breezeframework/dsl/json/JsonDslTest.kt
@@ -3,7 +3,7 @@ package com.windea.breezeframework.dsl.json
 import org.junit.*
 
 class JsonDslTest {
-	@Test //TESTED OK
+	@Test OK
 	fun test1() {
 		jsonTree {
 			listOf(
@@ -13,8 +13,8 @@ class JsonDslTest {
 			)
 		}.also { println(it) }
 	}
-	
-	@Test //TESTED OK
+
+	@Test OK
 	fun test2() {
 		json {
 			jsonArrayOf(
@@ -24,8 +24,8 @@ class JsonDslTest {
 			)
 		}.also { println(it) }
 	}
-	
-	@Test //TESTED OK
+
+	@Test OK
 	fun test3() {
 		json {
 			jsonArrayOf(
@@ -35,8 +35,8 @@ class JsonDslTest {
 			)
 		}.also { println(it) }
 	}
-	
-	@Test //TESTED OK
+
+	@Test OK
 	fun test4() {
 		JsonConfig.prettyPrint = false
 		jsonTree {
@@ -47,8 +47,8 @@ class JsonDslTest {
 			)
 		}.also { println(it) }
 	}
-	
-	@Test //TESTED OK
+
+	@Test OK
 	fun test5() {
 		jsonTree {
 			listOf(